+++ src/Humanizer.Tests/Localisation/DefaultFormatterTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Globalization;
+using System.Threading;
 using Humanizer.Localisation;
 using Humanizer.Localisation.Formatters;
 using Xunit;
@@ -19,6 +20,18 @@ namespace Humanizer.Tests.Localisation
         public void TimeSpanHumanizeThrowsExceptionForTimeUnitsLargerThanWeek(TimeUnit timeUnit, int unit)
         {
             Assert.Throws<ArgumentOutOfRangeException>(() => new DefaultFormatter(CultureInfo.InvariantCulture.Name).TimeSpanHumanize(timeUnit, unit));
+        }
+
+        [Fact]
+        public void HandlesNotImplementedCollectionFormattersGracefully()
+        {
+            using (new AmbientCulture("es"))
+            {
+                var a = new[] { DateTime.UtcNow, DateTime.UtcNow.AddDays(10) };
+                var b = a.Humanize(); 
+
+                Assert.Equal(a[0] + " & " + a[1], b);
+            }
         }
     }
 }