+++ release_notes.md
@@ -1,5 +1,7 @@
 ###In Development
 
+ - [#392](https://github.com/MehdiK/Humanizer/issues/392): Default implementation for collection formatter using regular style instead of an exception. Default separator is ampercent.
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.34.0...master)
 
 ###v1.34.0 - 2015-03-04