+++ release_notes.md
@@ -1,5 +1,6 @@
 ###In Development
- - [#399](https://github.com/MehdiK/Humanizer/pull/399): Support for humanizing DateTimeOffset
+ - [#399](https://github.com/MehdiK/Humanizer/pull/399): Added support for humanizing DateTimeOffset
+ - [#395](https://github.com/MehdiK/Humanizer/pull/395): Added support for Xamarin platforms to PCL 
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.34.0...master)