+++ release_notes.md
@@ -2,6 +2,8 @@
  - [#399](https://github.com/MehdiK/Humanizer/pull/399): Added support for humanizing DateTimeOffset
  - [#395](https://github.com/MehdiK/Humanizer/pull/395): Added support for Xamarin platforms to PCL 
 
+ - [#392](https://github.com/MehdiK/Humanizer/issues/392): Default implementation for collection formatter using regular style instead of an exception. Default separator is ampercent.
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.34.0...master)
 
 ###v1.34.0 - 2015-03-04