+++ release_notes.md
@@ -1,11 +1,15 @@
 ###In Development
+
+[Commits](https://github.com/MehdiK/Humanizer/compare/v1.35.0...master)
+
+###v1.35.0 - 2015-03-29
  - [#399](https://github.com/MehdiK/Humanizer/pull/399): Added support for humanizing DateTimeOffset
  - [#395](https://github.com/MehdiK/Humanizer/pull/395): Added support for Xamarin platforms to PCL 
  - [#397](https://github.com/MehdiK/Humanizer/pull/397): Added is/are to inflector rules
  - [#392](https://github.com/MehdiK/Humanizer/pull/392): Default implementation for collection formatter using regular style instead of an exception. Default separator is ampercent.
  - [#377](https://github.com/MehdiK/Humanizer/pull/392): Added culture specific decimal separator
 
-[Commits](https://github.com/MehdiK/Humanizer/compare/v1.34.0...master)
+[Commits](https://github.com/MehdiK/Humanizer/compare/v1.34.0...v1.35.0)
 
 ###v1.34.0 - 2015-03-04
  - [#381](https://github.com/MehdiK/Humanizer/pull/381): Fixes trailing question mark reported in #378.